Authenticate Requests

+++ todo_app/routers/todos.py
@@ -3,10 +3,14 @@ from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
 from fastapi import APIRouter, Depends, HTTPException, Path
 from starlette import status
-from todo_app.database import SessionLocal
-from todo_app.models import Todo
+from ..database import SessionLocal
+from ..models import Todo
+from .auth import get_current_user
 
-router = APIRouter()
+router = APIRouter(
+    prefix="/todo",
+    tags=["todo"],
+)
 
 
 def get_db():
@@ -18,6 +22,7 @@ def get_db():
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
@@ -26,24 +31,13 @@ class TodoRequest(BaseModel):
     priority: int = Field(gt=0)
     complete: bool
 
-    def model_dump(self, *args, **kwargs):
-        # Tùy chỉnh thứ tự các trường
-        data = super().model_dump(*args, **kwargs)
-        return {
-            "id": data["id"],
-            "title": data["title"],
-            "description": data["description"],
-            "priority": data["priority"],
-            "complete": data["complete"],
-        }
-
 
 @router.get("/")
-async def read_all(db: db_dependency):
-    return db.query(Todo).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    return db.query(Todo).filter(Todo.owner_id == user.get('id')).all()
 
 
-@router.get('/todo/{todo_id}', status_code=status.HTTP_200_OK)
+@router.get('/{todo_id}', status_code=status.HTTP_200_OK)
 async def read_one(db: db_dependency, todo_id: int = Path(gt=0)):
     todo_model = db.query(Todo).filter(Todo.id == todo_id).first()
     if todo_model is not None:
@@ -51,17 +45,27 @@ async def read_one(db: db_dependency, todo_id: int = Path(gt=0)):
     raise HTTPException(status_code=404, detail="Todo not found")
 
 
-@router.post('/todo', status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todo(**todo_request.model_dump())
+@router.post('/', status_code=status.HTTP_201_CREATED)
+async def create_todo(user: user_dependency,
+                      db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail='Authentication Failed.')
+    todo_model = Todo(**todo_request.dict(), owner_id=user.get('id'))
 
     db.add(todo_model)
     db.commit()
 
 
-@router.put('/todo/{todo_id}', status_code=status.HTTP_200_OK)
-async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todo).filter(Todo.id == todo_id).first()
+@router.put('/{todo_id}', status_code=status.HTTP_200_OK)
+async def update_todo(user: user_dependency, db: db_dependency,
+                      todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail='Authentication Failed.')
+
+    todo_model = (db.query(Todo).filter(Todo.id == todo_id)
+                  .filter(Todo.owner_id == user.get('id')).first())
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
     todo_model.title = todo_request.title
@@ -73,11 +77,16 @@ async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int
     db.commit()
 
 
-@router.delete('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todo).filter(Todo.id == todo_id).first()
+@router.delete('/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail='Authentication Failed.')
+    todo_model = (db.query(Todo).filter(Todo.id == todo_id).
+                  filter(Todo.owner_id == user.get('id')).first())
+
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
 
-    db.query(Todo).filter(Todo.id == todo_id).delete()
+    db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).delete()
     db.commit()
